Moved fn searchKeywords from proposals models to utils and refactored the fn into two separate ones
