+++ commercialoperator/components/proposals/models.py
@@ -6262,216 +6262,6 @@ def duplicate_event(p):
 
     return p
 
-def searchKeyWords(
-    context,
-    request,
-    searchWords,
-    searchProposal,
-    searchApproval,
-    searchCompliance,
-    is_internal=True,
-):
-    from commercialoperator.utils import (
-        search_approval,
-        search_compliance,
-        getChoiceFieldRegex,
-    )
-    from commercialoperator.components.approvals.models import Approval
-    from commercialoperator.components.compliances.models import Compliance
-    from commercialoperator.utils import search
-
-    application_types = [
-        ApplicationType.TCLASS,
-        ApplicationType.EVENT,
-        ApplicationType.FILMING,
-    ]
-
-    proposal_list = Proposal.objects.none()
-    approval_list =Approval.objects.none()
-    compliance_list = Compliance.objects.none()
-
-    if is_internal and searchProposal:
-        proposal_list = Proposal.objects.filter(
-            application_type__name__in=application_types
-        ).exclude(processing_status__in=["discarded", "draft"])
-    if is_internal and searchApproval:
-        approval_list = (
-            Approval.objects.all()
-            .order_by("lodgement_number", "-issue_date")
-            .distinct("lodgement_number")
-        )
-    if is_internal and searchCompliance:
-        compliance_list = Compliance.objects.all()
-
-    if searchWords:
-        # convert the search words in to two regex values - one for text one for json values
-        search_words_regex = "(?:" + "|".join(searchWords) + ")"
-        filter_regex = (
-            '.*".*":\s"(\\\\"|[^"])*' + search_words_regex + '(\\\\"|[^"])*".*'
-        )
-
-        # three searchable fields use choices: accreditation, film_type, and film_purpose
-        # so we must convert the search phrase in to the search short word where applicable
-        # accreditation
-        accreditation_words = getChoiceFieldRegex(
-            searchWords, ProposalAccreditation.ACCREDITATION_TYPE_CHOICES
-        )
-        # film type
-        film_type_words = getChoiceFieldRegex(
-            searchWords, ProposalFilmingActivity.FILM_TYPE_CHOICES
-        )
-        # film purpose
-        film_purpose_words = getChoiceFieldRegex(
-            searchWords, ProposalFilmingActivity.PURPOSE_CHOICES
-        )
-
-        # one particular search value is quite nested - it is faster to run these queries instead of the reverse
-        activities = Activity.objects.filter(name__iregex=search_words_regex)
-        pts_activities = ProposalTrailSectionActivity.objects.filter(
-            activity__in=activities
-        )
-        sections = ProposalTrailSection.objects.filter(
-            trail_activities__in=pts_activities
-        )
-        trails = ProposalTrail.objects.filter(sections__in=sections)
-
-        paginator = context.paginator
-        paginator.page_size = 10
-
-        if searchProposal:
-            # this below query run is equivalent to the search_words property, except that it retrieves the pertaining proposal records
-            # there is a lot here but a lot of time is saved not having to iterate every record every time
-            proposal_list = (
-                proposal_list.filter(
-                    (
-                        Q(application_type__name=ApplicationType.TCLASS)
-                        & (Q(parks__park__name__iregex=search_words_regex))
-                        | (Q(parks__activities__activity__in=activities))
-                        | (Q(parks__zones__park_activities__activity__in=activities))
-                        | (Q(trails__trail__name__iregex=search_words_regex))
-                        | (Q(vehicles__rego__iregex=search_words_regex))
-                        | (Q(vessels__spv_no__iregex=search_words_regex))
-                        | (Q(other_details__other_comments__iregex=search_words_regex))
-                        | (Q(other_details__mooring__iregex=search_words_regex))
-                        | (
-                            Q(
-                                other_details__accreditations__accreditation_type__in=accreditation_words
-                            )
-                        )
-                        | (Q(trails__in=trails))
-                    )
-                    | (
-                        Q(application_type__name=ApplicationType.EVENT)
-                        & (Q(events_parks__park__name__iregex=search_words_regex))
-                        | (Q(events_parks__event_activities__iregex=search_words_regex))
-                        | (Q(trails__trail__name__iregex=search_words_regex))
-                        | (Q(vehicles__rego__iregex=search_words_regex))
-                        | (Q(vessels__spv_no__iregex=search_words_regex))
-                        | (Q(trails__in=trails))
-                    )
-                    | (
-                        Q(application_type__name=ApplicationType.FILMING)
-                        & (Q(filming_parks__park__name__iregex=search_words_regex))
-                        | (Q(vehicles__rego__iregex=search_words_regex))
-                        | (Q(vessels__spv_no__iregex=search_words_regex))
-                        | (Q(filming_activity__film_type__in=film_type_words))
-                        | (Q(filming_activity__film_purpose__in=film_purpose_words))
-                    )
-                )
-                .distinct("id")
-                .order_by("-id")
-            )
-
-        if searchApproval:
-            approval_list = approval_list.filter(
-                Q(surrender_details__iregex=filter_regex)
-                | Q(suspension_details__iregex=filter_regex)
-                | Q(cancellation_details__iregex=search_words_regex)
-            )
-
-        if searchCompliance:
-            compliance_list = compliance_list.filter(
-                Q(text__iregex=search_words_regex)
-                | Q(requirement__free_requirement__iregex=search_words_regex)
-                | Q(requirement__standard_requirement__text__iregex=search_words_regex)
-            )
-
-    return_list = []
-    chained_list = QuerySetChain(proposal_list, approval_list, compliance_list)
-    chained_list_paginated = paginator.paginate_queryset(chained_list, request)
-
-    for entry in chained_list_paginated:
-        if isinstance(entry, Proposal):
-            if not entry.search_data:
-                continue
-
-            search_results = search(entry.search_data, searchWords)
-            if not len(search_results):
-                continue
-            search_results = json.dumps(search(entry.search_data, searchWords))
-
-            search_result_tuple = (entry.id, entry.lodgement_number, json.loads(search_results))
-
-            final_results = {}
-            pid = search_result_tuple[0]
-            lodgement_number = search_result_tuple[1]
-
-            for result in search_result_tuple[2]:
-                for key, value in result.items():
-                    final_results.update({"key": key, "value": value})
-
-            cache_key = settings.CACHE_KEY_PROPOSAL_KEYWORD_SEARCH.format(
-                id=pid, lodgement_number=lodgement_number
-            )
-            res = cache.get(cache_key)
-
-            if res is None:
-                try:
-                    applicant = proposal_list.get(id=pid).applicant_obj
-                except Proposal.DoesNotExist:
-                    applicant = None
-
-                res = {
-                    "number": lodgement_number,
-                    "id": pid,  # id,
-                    "type": "Proposal",
-                    "applicant": applicant,
-                }
-
-                cache.set(
-                    cache_key,
-                    res,
-                    settings.CACHE_TIMEOUT_24_HOURS,
-                )
-            else:
-                logger.info(
-                    "Search Keywords cache hit for proposal {}".format(
-                        res["number"]
-                    )
-                )
-
-            res["text"] = final_results
-
-            return_list.append(res)
-
-        elif isinstance(entry, Approval):
-            try:
-                results = search_approval(entry, searchWords)
-                return_list.extend(results)
-            except:
-                pass
-        elif isinstance(entry, Compliance):
-            try:
-                results = search_compliance(entry, searchWords)
-                return_list.extend(results)
-            except:
-                pass
-        else:
-            raise ValueError(f"Unknown entry type {type(entry)} in search results {entry}")
-
-    return return_list
-
-
 def search_reference(reference_number):
     from commercialoperator.components.approvals.models import Approval
     from commercialoperator.components.compliances.models import Compliance
